+++ 08.ls_object/bin/ls.rb
@@ -4,11 +4,13 @@
 require 'optparse'
 require 'io/console'
 require_relative '../lib/ls'
+require_relative '../lib/short_format'
+require_relative '../lib/long_format'
 require_relative '../lib/file'
 
 def main
   ls = Command::Ls.new(ARGV.getopts('arl'), IO.console.winsize[1])
-  ls.execute.inject { |res, str| res + str }
+  print(ls.execute.inject { |res, str| res + str })
 end
 
 main
